+++ genesyscloud/telephony_providers_edges_edge_group/data_source_genesyscloud_telephony_providers_edges_edge_group.go
@@ -14,26 +14,6 @@ import (
 	"github.com/mypurecloud/platform-client-sdk-go/v116/platformclientv2"
 )
 
-func DataSourceEdgeGroup() *schema.Resource {
-	return &schema.Resource{
-		Description: "Data source for Genesys Cloud Edge Group. Select an edge group by name",
-		ReadContext: gcloud.ReadWithPooledClient(dataSourceEdgeGroupRead),
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Description: "Edge Group name.",
-				Type:        schema.TypeString,
-				Required:    true,
-			},
-			"managed": {
-				Description: "Return entities that are managed by Genesys Cloud.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-				Default:     false,
-			},
-		},
-	}
-}
-
 func dataSourceEdgeGroupRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	sdkConfig := m.(*gcloud.ProviderMeta).ClientConfig
 	edgesAPI := platformclientv2.NewTelephonyProvidersEdgeApiWithConfig(sdkConfig)
