+++ genesyscloud/telephony_providers_edges_edge_group/resource_genesyscloud_telephony_providers_edges_edge_group.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"strings"
 	gcloud "terraform-provider-genesyscloud/genesyscloud"
 	"time"
 
@@ -13,63 +12,11 @@ import (
 	"terraform-provider-genesyscloud/genesyscloud/consistency_checker"
 
 	resourceExporter "terraform-provider-genesyscloud/genesyscloud/resource_exporter"
-	lists "terraform-provider-genesyscloud/genesyscloud/util/lists"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/mypurecloud/platform-client-sdk-go/v116/platformclientv2"
 )
-
-func ResourceEdgeGroup() *schema.Resource {
-	return &schema.Resource{
-		Description: `Genesys Cloud Edge Group. NOTE: This resource is being kept here for backwards compatibility with older Genesys Cloud Organization. You may get an error if you try to create an edge group with a Genesys Cloud Organization created in 2022 or later.`,
-
-		CreateContext: gcloud.CreateWithPooledClient(createEdgeGroup),
-		ReadContext:   gcloud.ReadWithPooledClient(readEdgeGroup),
-		UpdateContext: gcloud.UpdateWithPooledClient(updateEdgeGroup),
-		DeleteContext: gcloud.DeleteWithPooledClient(deleteEdgeGroup),
-		Importer: &schema.ResourceImporter{
-			StateContext: schema.ImportStatePassthroughContext,
-		},
-		SchemaVersion: 1,
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Description: "The name of the entity.",
-				Type:        schema.TypeString,
-				Required:    true,
-			},
-			"description": {
-				Description: "The resource's description.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"state": {
-				Description: "Indicates if the resource is active, inactive, or deleted.",
-				Type:        schema.TypeString,
-				Optional:    true,
-				Computed:    true,
-			},
-			"managed": {
-				Description: "Is this edge group being managed remotely.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-				Default:     false,
-			},
-			"hybrid": {
-				Description: "Is this edge group hybrid.",
-				Type:        schema.TypeBool,
-				Optional:    true,
-				Default:     false,
-			},
-			"phone_trunk_base_ids": {
-				Description: "A list of trunk base settings IDs of trunkType \"PHONE\" to inherit to edge logical interface for phone communication.",
-				Type:        schema.TypeSet,
-				Required:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-			},
-		},
-	}
-}
 
 func createEdgeGroup(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	name := d.Get("name").(string)
@@ -216,21 +163,13 @@ func readEdgeGroup(ctx context.Context, d *schema.ResourceData, meta interface{}
 		}
 		d.Set("phone_trunk_base_ids", nil)
 		if edgeGroup.PhoneTrunkBases != nil {
-			d.Set("phone_trunk_base_ids", flattenPhoneTrunkBases(*edgeGroup.PhoneTrunkBases))
+			d.Set("phone_trunk_base_ids", FlattenPhoneTrunkBases(*edgeGroup.PhoneTrunkBases))
 		}
 
 		log.Printf("Read edge group %s %s", d.Id(), *edgeGroup.Name)
 
 		return cc.CheckState()
 	})
-}
-
-func flattenPhoneTrunkBases(trunkBases []platformclientv2.Trunkbase) *schema.Set {
-	interfaceList := make([]interface{}, len(trunkBases))
-	for i, v := range trunkBases {
-		interfaceList[i] = *v.Id
-	}
-	return schema.NewSet(schema.HashString, interfaceList)
 }
 
 func getAllEdgeGroups(_ context.Context, sdkConfig *platformclientv2.Configuration) (resourceExporter.ResourceIDMetaMap, diag.Diagnostics) {
@@ -275,51 +214,4 @@ func getAllEdgeGroups(_ context.Context, sdkConfig *platformclientv2.Configurati
 	}
 
 	return resources, nil
-}
-
-func EdgeGroupExporter() *resourceExporter.ResourceExporter {
-	return &resourceExporter.ResourceExporter{
-		GetResourcesFunc: gcloud.GetAllWithPooledClient(getAllEdgeGroups),
-		RefAttrs: map[string]*resourceExporter.RefAttrSettings{
-			"phone_trunk_base_ids": {RefType: "genesyscloud_telephony_providers_edges_trunkbasesettings"},
-		},
-	}
-}
-
-func buildSdkTrunkBases(d *schema.ResourceData) *[]platformclientv2.Trunkbase {
-	returnValue := make([]platformclientv2.Trunkbase, 0)
-
-	if ids, ok := d.GetOk("phone_trunk_base_ids"); ok {
-		phoneTrunkBaseIds := lists.SetToStringList(ids.(*schema.Set))
-		for _, trunkBaseId := range *phoneTrunkBaseIds {
-			id := trunkBaseId
-			returnValue = append(returnValue, platformclientv2.Trunkbase{
-				Id: &id,
-			})
-		}
-	}
-
-	return &returnValue
-}
-
-func GenerateEdgeGroupResourceWithCustomAttrs(
-	edgeGroupRes,
-	name,
-	description string,
-	managed,
-	hybrid bool,
-	otherAttrs ...string) string {
-	return fmt.Sprintf(`resource "genesyscloud_telephony_providers_edges_edge_group" "%s" {
-		name = "%s"
-		description = "%s"
-		managed = "%v"
-		hybrid = "%v"
-		%s
-	}
-	`, edgeGroupRes, name, description, managed, hybrid, strings.Join(otherAttrs, "\n"))
-}
-
-func GeneratePhoneTrunkBaseIds(userIDs ...string) string {
-	return fmt.Sprintf(`phone_trunk_base_ids = [%s]
-	`, strings.Join(userIDs, ","))
 }
